+++ docs/conf.py
@@ -192,9 +192,6 @@ nitpick_ignore = [
     ("py:class", "substra.sdk.schemas.AlgoOutputSpec"),
     ("py:class", "substra.sdk.schemas.AlgoInputSpec"),
     ("py:class", "ComputePlan"),
-    ("py:class", "substratools.algo.GenericAlgo"),
-    ("py:class", "substratools.algo.CompositeAlgo"),
-    ("py:class", "substratools.algo.AggregateAlgo"),
 ]
 
 html_css_files = [