+++ substrafl/nodes/test_data_node.py
@@ -133,7 +133,7 @@ class TestDataNode(Node):
         dependencies: Dependency,
     ) -> Dict[RemoteStruct, OperationKey]:
         """Find the algorithms from the parent traintuple of each predicttuple and submit it with a dockerfile
-        specifying the ``predict`` method as ``--method-name`` to execute.
+        specifying the ``predict`` method as ``--function-name`` to execute.
 
         Go through every operation in the predict algo cache, submit it to substra and save `RemoteStruct : algo_key`
         into the `cache` (where algo_key is the returned algo key by substra.)